requirements.txt
streamlit
pandas
spacy
joblib
numpy
scikit-learn
PyPDF2
python-docx
fpdf
matplotlib
seaborn
